requirements.txt
PySide6>=6.5.0
openai>=1.0.0
Pillow>=9.0.0
pygetwindow>=0.0.9
pyautogui>=0.9.54
dotenv